Restore arm types to their original settings
These match GCCs settings.

+++ arch/arm/include/_types/_intptr_t.h
@@ -1,6 +1,6 @@
 #ifndef __INTPTR_T_H_
 #define __INTPTR_T_H_
 
-typedef long int intptr_t;
+typedef int intptr_t;
 
 #endif // __INTPTR_T_H_
